+++ test/features/field.js
@@ -273,11 +273,25 @@ define(function () {
         expect($('<input type="email" value="">').parsley().isValid(true)).to.be(false);
         expect($('<input type="email" value="foo">').parsley().isValid(true, "")).to.be(false);
       });
+      it('should have a whitespace="squish" option', function () {
+        $('body').append('<input type="text" id="element" value=" foo    bar " />');
+        expect($('#element').parsley().getValue()).to.be(' foo    bar ');
+        $('#element').attr('data-parsley-whitespace', 'squish').parsley().actualizeOptions();
+        expect($('#element').parsley().getValue()).to.be('foo bar');
+      });
+      it('should have a whitespace="trim" option', function () {
+        $('body').append('<input type="text" id="element" value=" foo " />');
+        expect($('#element').parsley().getValue()).to.be(' foo ');
+        $('#element').attr('data-parsley-whitespace', 'trim').parsley().actualizeOptions();
+        expect($('#element').parsley().getValue()).to.be('foo');
+      });
       it('should have a trim-value option', function () {
         $('body').append('<input type="text" id="element" value=" foo " />');
         expect($('#element').parsley().getValue()).to.be(' foo ');
         $('#element').attr('data-parsley-trim-value', true).parsley().actualizeOptions();
-        expect($('#element').parsley().getValue()).to.be('foo');
+        expectWarning(function() {
+          expect($('#element').parsley().getValue()).to.be('foo');
+        });
       });
       it('should inherit options from the form, even if the form is bound after', function () {
         $('body').append('<form id="element" data-parsley-required>' +